Remove crates that have been deleted from crates.io
These crates were removed for abuse. Since the crates were removed from
the crates.io database, one name has since had a new crate published.
For that crate (deposit), the file is still present, with only the
deleted version being removed.

Due to human error, a few of the crates that were removed from crates.io
were not documented at the time, so there may still be a small number of
crates present in the index which are not present in crates.io. This
will be resolved in the future.