Sync index with crates.io database
These crates/versions are present in the index, but are not present in
the crates.io database. `std` and `rustbook` were likely removed from
the database manually when we added those to reserved crate names on our
end, but were never removed from the index.

The missing version of cargo-ctags was most likely the result of a rare
situation where the index is updated, but the server shuts down before
the database transaction is committed.

I have no idea what's up with `ojfiewijogwhiogerhiugerhiuegr`, but it's
a random enough name that I'm willing to accept not knowing what
happened there.

This list was found by reading every file in the index and ensuring that
every version/name pair was present in the crates.io database. After
this commit, there are no more cases of any crates or versions in the
index that are not in the crates.io database.